sun radians represented by double instead of float

+++ include/scene.hpp
@@ -12,7 +12,7 @@ namespace lowpoly3d {
 struct Scene {
 	std::vector<RenderData> renderDatas;
 	glm::mat4 view;
-	float sunRadians;
+	double sunRadians;
 	bool showWireframes;
 };
